use TimePrecision for precision literal fields

+++ literal/utils.go
@@ -333,7 +333,7 @@ func NewDecimalFromApdDecimal(value *apd.Decimal, nullable bool) (expr.Literal, 
 
 func NewPrecisionTime(precision types.TimePrecision, value int64, nullable bool) (expr.Literal, error) {
 	return expr.NewLiteral(&types.PrecisionTime{
-		Precision: int32(precision),
+		Precision: precision,
 		Value:     value,
 	}, nullable)
 }
@@ -362,7 +362,7 @@ func NewPrecisionTimestampFromTime(precision types.TimePrecision, tm time.Time, 
 // NewPrecisionTimestamp creates a new PrecisionTimestamp literal with given precision and value.
 func NewPrecisionTimestamp(precision types.TimePrecision, value int64, nullable bool) (expr.Literal, error) {
 	return expr.NewLiteral(&types.PrecisionTimestamp{
-		Precision: int32(precision),
+		Precision: precision,
 		Value:     value,
 	}, nullable)
 }
@@ -383,7 +383,7 @@ func NewPrecisionTimestampTzFromTime(precision types.TimePrecision, tm time.Time
 // NewPrecisionTimestampTz creates a new PrecisionTimestampTz literal with given precision and value.
 func NewPrecisionTimestampTz(precision types.TimePrecision, value int64, nullable bool) (expr.Literal, error) {
 	return expr.NewLiteral(&types.PrecisionTimestampTz{
-		Precision: int32(precision),
+		Precision: precision,
 		Value:     value,
 	}, nullable)
 }
